requirements.txt
flask
python-dotenv
requests
SQLAlchemy
Flask-SQLAlchemy
werkzeug